events validates added and function to diside changed

+++ app/controllers/events_controller.rb
@@ -51,7 +51,6 @@ class EventsController < ApplicationController
     @event = Event.new
 
     if @act == 'group'
-      # @event.timeplans.build
       2.times { @event.timeplans.build }
     end
     
@@ -85,8 +84,6 @@ class EventsController < ApplicationController
     else
       @event.users << current_user
 
-      # current_user.events.create(event_params)
-
     end
     respond_to do |format|
       if @event.save
@@ -109,25 +106,30 @@ class EventsController < ApplicationController
   def deside
     # 確定したプランを取得
     tp = Timeplan.find(params[:tp_id])
-    @event.transaction do
-      # 開始と終了とカラーを更新
-      @event.update(start: tp.start, end: tp.end, color: 'green')
-      # イベント情報を更新
-      @event.update(event_params)
+    if Entry.where(timeplan_id: tp.id, attendance: true).count == 0
+      redirect_to @event, alert: '参加者がいないイベントは登録できません'
 
-      # 参加できるユーザーを登録
-      tp.entries.each do |ent|
-        if ent.attendance == true
-          @event.users << ent.user
+    else
+      @event.transaction do
+        # 開始と終了とカラーを更新
+        @event.update(start: tp.start, end: tp.end, color: 'green')
+        # イベント情報を更新
+        @event.update(event_params)
+
+        # 参加できるユーザーを登録
+        tp.entries.each do |ent|
+          if ent.attendance == true
+            @event.users << ent.user
+          end
         end
+        # イベントに関連するtimeplansレコードとentriesレコードを削除
+        @event.timeplans.each do |tp2|
+          tp2.entries.destroy_all
+          tp2.destroy
+        end
+        # raise '例外'
+         redirect_to @event
       end
-      # イベントに関連するtimeplansレコードとentriesレコードを削除
-      @event.timeplans.each do |tp2|
-        tp2.entries.destroy_all
-        tp2.destroy
-      end
-      # raise '例外'
-      redirect_to @event
     end
   rescue => e
     render text: e.message
